+++ tests/testthat/test-GeoTox-package.R
@@ -1,3 +1,3 @@
 test_that("ignore imports", {
-  expect_silent(ignore_unused_imports())
+  expect_no_error(ignore_unused_imports())
 })